[KERNEL32][SMSS][TASKMGR] Compute physical memory sizes in 64 bits
NumberOfPhysicalPages and PageSize are both ULONG, so multiplying them
evaluated in 32 bits and wrapped before the result was widened.  On a machine
with 8 GB of RAM GlobalMemoryStatusEx() reported 4193732 kB, making it look as
if the kernel had capped physical memory at 4 GB when it had in fact detected
all of it.

Widen the multiplications:
- kernel32: ullTotalPhys and ullAvailPhys.
- smss: the system-managed pagefile is sized from RAM, so the wrap picked
  minimum and maximum sizes for a machine several gigabytes smaller.
- taskmgr: the system cache figure is converted to kilobytes before the
  multiply, matching the neighbouring accessors.

GlobalMemoryStatus() needs no change: it already saturates the 32-bit fields
with min(..., MAXULONG_PTR) after calling the extended version.

+++ base/system/smss/pagefile.c
@@ -732,7 +732,7 @@ SmpMakeSystemManagedPagingFileDescriptor(IN PSMP_PAGEFILE_DESCRIPTOR Descriptor)
     }
 
     /* Check how much RAM we have and set three times this amount as maximum */
-    Ram = BasicInfo.NumberOfPhysicalPages * BasicInfo.PageSize;
+    Ram = (ULONGLONG)BasicInfo.NumberOfPhysicalPages * BasicInfo.PageSize;
     MaximumSize = 3 * Ram;
 
     /* If we have more than 1GB, use that as minimum, otherwise, use 1.5X RAM */
